Wrap CloudFlare client. Add zone id caching to start
